Change low-level MIDI sendImpl functions Support sending chunked SysEx using BLE and USBMIDI Change SysEx chunk size to uint16_t (instead of uint8_t)

+++ src/MIDI_Parsers/SysExBuffer.cpp
@@ -33,6 +33,6 @@ bool SysExBuffer::isReceiving() const { return receiving; }
 
 const uint8_t *SysExBuffer::getBuffer() const { return buffer; }
 
-size_t SysExBuffer::getLength() const { return length; }
+uint16_t SysExBuffer::getLength() const { return length; }
 
 END_CS_NAMESPACE
